# openapi/schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Karag API",
    "description": "Modular RAG & Agentic Chatbot API",
    "contact": {
      "name": "Karag Team",
      "url": "https://github.com/karag-ai/karag"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/v1/documents/{document_id}": {
      "get": {
        "tags": [
          "openai"
        ],
        "summary": "Get Document For Citation",
        "description": "Retrieve a document by ID for citation reverse lookup.\n\nThis endpoint allows clients to resolve [[doc:<document_id>]] citations\nfrom chat completions. It returns the document metadata and content preview.\n\nRequires authentication and workspace access authorization.\n\nArgs:\n    document_id: The unique document identifier from a citation\n    workspace_id: Required workspace ID for access validation\n    user: Authenticated user context\n\nReturns:\n    DocumentCitationResponse with metadata and content preview\n\nRaises:\n    HTTPException: If user lacks workspace access or document not found",
        "operationId": "get_document_for_citation_api_v1_v1_documents__document_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentCitationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/workspaces/{workspace_id}/documents/{document_id}": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get Document",
        "operationId": "get_document_api_v1_workspaces__workspace_id__documents__document_id__get",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_DocumentResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Document",
        "operationId": "delete_document_api_v1_workspaces__workspace_id__documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          },
          {
            "name": "workspace_id",
            "in": "path",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Workspace Id"
            }
          },
          {
            "name": "dataset_delete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dataset Delete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppResponse_dict_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppResponse_DocumentResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocumentResponse"
              },
              {
                "type": "null"
              }
            ],
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[DocumentResponse]"
      },
...
      "AppResponse_dict_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Whether the operation was successful",
            "default": true
          },
          "code": {
            "type": "string",
            "title": "Code",
            "description": "Machine-readable status code",
            "default": "SUCCESS"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "default": "Operation completed successfully"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data",
            "description": "Response payload (type varies by endpoint)"
          }
        },
        "type": "object",
        "title": "AppResponse[dict]"
      },
...
      "DocumentCitationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Document ID from citation"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "MIME type"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "description": "Owning workspace"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source information"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Processing status"
          },
          "content_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Preview",
            "description": "First 5000 chars of content"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "filename",
          "content_type",
          "workspace_id",
          "status"
        ],
        "title": "DocumentCitationResponse",
        "description": "Response for citation document lookup.\n\nUsed when frontend resolves [[doc:<document_id>]] from chat responses.",
        "example": {
          "content_preview": "Abstract: This paper discusses...",
          "content_type": "application/pdf",
          "filename": "research_paper.pdf",
          "id": "doc_abc123",
          "source": "upload",
          "status": "indexed",
          "workspace_id": "ws_xyz789"
        }
      },
...
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique document ID (stable, persistent)"
          },
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id",
            "description": "Workspace this document belongs to"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "MIME type"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source URL or description"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Processing status",
            "default": "uploaded"
          },
          "size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size",
            "description": "File size in bytes"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks",
            "description": "Number of chunks indexed",
            "default": 0
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content",
            "description": "Document content (truncated for preview)"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url",
            "description": "Presigned URL for download"
          }
        },
        "type": "object",
        "required": [
          "id",
          "workspace_id",
          "filename",
          "content_type"
        ],
        "title": "DocumentResponse",
        "description": "Standard document response for frontend integration.\n\nThis schema is used for:\n- Document listing in workspace storage\n- Citation click-through from chat\n- Document metadata display",
        "example": {
          "chunks": 42,
          "content_type": "application/pdf",
          "created_at": "2024-01-15T10:30:00Z",
          "filename": "document.pdf",
          "id": "doc_abc123",
          "size": 1024000,
          "source": "upload",
          "status": "indexed",
          "updated_at": "2024-01-15T10:35:00Z",
          "workspace_id": "ws_xyz789"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}